Fix some Linux issues.
git-svn-id: https://cvs.khronos.org/svn/repos/ogl/trunk/ecosystem/public/sdk/tools/glslang@23939 e7fa87d3-cd2b-0410-9028-fcbf551c1848

+++ glslang/MachineIndependent/reflection.cpp
@@ -312,14 +312,14 @@ void TReflection::dump()
 {
     printf("Uniform reflection:\n");
     for (size_t i = 0; i < indexToUniform.size(); ++i) {
-        printf("%d:", i);
+        printf("%d:", (int)i);
         indexToUniform[i].dump();
     }
     printf("\n");
 
     printf("Uniform block reflection:\n");
     for (size_t i = 0; i < indexToUniformBlock.size(); ++i) {
-        printf("%d: ", i);
+        printf("%d: ", (int)i);
         indexToUniformBlock[i].dump();
     }
     printf("\n");
